+++ packages/sku/src/services/vite/index.ts
@@ -1,4 +1,5 @@
 import { build, createServer } from 'vite';
+
 import type { SkuContext } from '../../context/createSkuContext.js';
 
 import {
@@ -9,9 +10,9 @@ import {
 import { cleanTargetDirectory } from '../../utils/buildFileUtils.js';
 import { createOutDir } from './helpers/bundleConfig.js';
 import { getAppHosts } from '../../context/hosts.js';
-import chalk from 'chalk';
 import { prerenderConcurrently } from './helpers/prerender/prerenderConcurrently.js';
 import allocatePort from '../../utils/allocatePort.js';
+import { printUrls } from '@sku-lib/utils';
 
 export const viteService = {
   build: async (skuContext: SkuContext) => {
@@ -29,31 +30,18 @@ export const viteService = {
 
     const availablePort = await allocatePort({
       port: skuContext.port.client,
-      host: '0.0.0.0',
       strictPort: skuContext.port.strictPort,
     });
 
     await server.listen(availablePort);
 
     const hosts = getAppHosts(skuContext);
-    printUrls(hosts, skuContext);
+    printUrls(skuContext.listUrls ? hosts : [hosts[0]], {
+      https: skuContext.httpsDevServer,
+      initialPath: skuContext.initialPath,
+      port: availablePort,
+    });
 
     server.bindCLIShortcuts({ print: true });
   },
-};
-
-const printUrls = (
-  hosts: Array<string | undefined>,
-  skuContext: SkuContext,
-) => {
-  const proto = skuContext.httpsDevServer ? 'https' : 'http';
-  console.log('Starting development server...');
-  hosts.forEach((site) => {
-    const initialPath =
-      skuContext.initialPath !== '/' ? skuContext.initialPath : '';
-    const url = chalk.cyan(
-      `${proto}://${site}:${chalk.bold(skuContext.port.client)}${initialPath}`,
-    );
-    console.log(`${chalk.green('➜')}  ${chalk.bold('Local')}: ${url}`);
-  });
 };