+++ packages/utils/src/console/index.ts
@@ -1,2 +1,3 @@
 export { banner } from './banner.js';
 export { prependWithManagedConfigBanner } from './managedConfigBanner.js';
+export { printUrls } from './printUrls.js';